+++ docs/scripts/views/drawer/permanent-drawer-above-toolbar.vue
@@ -8,11 +8,16 @@
       </ui-drawer-header>
       <ui-drawer-content>
         <ui-nav>
-          <ui-nav-item :class="$theme.getTextClass('primary', $store.theme)" activated>Item {{ 0 }}</ui-nav-item>
           <ui-nav-item
+            href="javascript:void(0)"
+            active
             :class="$theme.getTextClass('primary', $store.theme)"
+          >Item {{ 0 }}</ui-nav-item>
+          <ui-nav-item
             v-for="i in 12"
             :key="i"
+            href="javascript:void(0)"
+            :class="$theme.getTextClass('primary', $store.theme)"
           >Item {{ i }}</ui-nav-item>
         </ui-nav>
       </ui-drawer-content>