+++ docs/scripts/views/drawer/permanent-drawer-below-toolbar.vue
@@ -18,12 +18,14 @@
         <ui-drawer-content>
           <ui-nav>
             <ui-nav-item
+              href="javascript:void(0)"
+              active
               :class="$theme.getTextClass('primary', $store.theme)"
-              activated
             >Item {{ 0 }}</ui-nav-item>
             <ui-nav-item
               v-for="i in 12"
               :key="i"
+              href="javascript:void(0)"
               :class="$theme.getTextClass('primary', $store.theme)"
             >Item {{ i }}</ui-nav-item>
           </ui-nav>